requirements.txt
streamlit
pandas
matplotlib
seaborn
langchain
langchain-google-genai
python-dotenv
